+++ macro/hvenh.py
@@ -8,7 +8,7 @@ import keyboard as shortcut
 from pynput.keyboard import Key, Listener, Controller
 from pynput import keyboard
 
-import util
+import common.util as util
 pynboard = Controller()
 
 # GLOBAL VARIABLES
@@ -32,10 +32,10 @@ def initialize(frame, btn, diff, runs=1):
   difficulty = diff
 
   shortcut.add_hotkey(util.HOTKEY_TERMINATE, util.terminate)
-  btn_start.config(state="disabled")
+  btn_start.config(state=util.STATE_DISABLED)
   frame_root.update()
   run_dungeon(runs)
-  btn_start.config(state="active")
+  btn_start.config(state=util.STATE_NORMAL)
   frame_root.update()
 
 def path_find(unit):